+++ changelog.md
@@ -5,6 +5,8 @@ FORMAT: 1A
 ## Release v1.0.1.26
 --------------------
 * FIX: Resolve deadlock on create/update where newly recognized acm being inserted by multiple concurrent transactions
+* CFG: Optional setting for deadlock retries can be configured via OD_DEADLOCK_RETRYCOUNTER, default is 5
+* CFG: Optional settinf for deadlock retry delay can be configured via OD_DEADLOCK_RETRYDELAYMS, default is 333.
 
 ## Release v1.0.1.25
 --------------------